+++ src/main/java/com/iseplife/api/dto/media/view/ImageView.java
@@ -6,7 +6,7 @@ import lombok.EqualsAndHashCode;
 @Data
 @EqualsAndHashCode(callSuper = true)
 public class ImageView extends MediaView {
-  private Double ratio;
+  private Float ratio;
   private String color;
   private Long thread;
 }
